*03* Add new Enemies that have Bullet
make 2 new classes :
1. EnemyBullet for bullet of enemies
2. ShootEnemy for body of new Enemy

And now heartScore is very OK!!!!

+++ f2/spw/GameReporter.java
@@ -3,6 +3,6 @@ package f2.spw;
 public interface GameReporter {
 
 	long getScore();
-	long getHeartScore();
+	int getHeartScore();
 
 }
